server/requirements.txt
fastapi==0.109.2
numpy==1.26.4
Pillow==10.1.0
Pillow==10.2.0
pydantic==2.6.1